Use non-blocking pipes and add debug envvar
Use non-blocking pipes for communications. This massively improves the
new-tab animation speed and stops the UI from blocking if the headless
process doesn't get started.

Add an environment variable, 'CLUTTER_MOZEMBED_DEBUG', that when
enabled, doesn't start up the headless process, but outputs the command
so you can run it yourself (in gdb, for example).

Also removed a lot of debug spew (not sure if that's all of it or not).
Eventually, there should be a global debugging env-var, or possibly
command-line option.