Split out view code to allow multiple views
Add a structure that holds all the data required for a view on the
headless process and store a list of these instead of just storing
single copies of the variables. We should be able to do multiple
viewports on a single head now (though only the first view will be
allowed full input/size control).