Add live out-of-process previews
Add a function that gets read-only ClutterMozEmbed actors that mirror
(live) any other open ClutterMozEmbed's in other applications. This
isn't incredibly reliable yet, but basically works. This is *VERY* hacky
at the moment and should be replaced with a nice dbus interface or
somesuch.

Also, the headless process will always try to remove input/output pipe
files, regardless of whether the pipe is open or not (less chance of
leaving pipes around, which tends to break the above).