Pulls in responsability for sizing/positioning the plugin viewport window
To simplify changes to the moblin-web-browser component clutter-mozembed
has now taken full ownership and responsability for the plugin viewport
window.

The window gets resized/positioned on actor::allocate, and the window
is moved off stage when the actor is hidden. (It's not unmapped since
that would conflict with the xcomposite redirection)