Fixes some of the plugin input handling issues involving multiple tabs
The scheme we now use is to move the plugin viewport window offscreen when the
corresponding tab is hidden, and back onscreen during actor::allocate when
the actor becomes visible.

There are still issues handling input while showing the tab previews, since
we also need to move the plugin windows offscreen while a mozembed actor
is un-reactive, but that will be dealt with that in a seperate patch.