Reparent plugin windows before destroying the viewport + add X error trapping
To try and avoid upsetting clutter-mozheadless we reparent plugin windows to
the stage (offscreen) before we destroy the plugin viewport. (during a dispose)
This means we don't also destroy the plugin windows from under the nose of
mozheadless.

We now also trap X errors when dealing with plugin windows since we can't be
sure they haven't been asynchronously destroyed resulting in BadWindow errors.