Remove horrible live preview code
Replaced the get_live_previews with two functions;
clutter_mozembed_new_view () and clutter_mozembed_connect_view (). It is
expected for the application to use whatever IPC to coordinate the use
of these two functions, where a view needs to use new_view and the host
application needs to call connect_view with the input and output
properties of the view actor.