Add connection timeouts + initial polling phase
Add a connection timeout to both ClutterMozEmbed and ClutterMozHeadless
at a default of 10 seconds, so that it gives up if a connection is never
made. This stops ClutterMozHeadless from doing nothing eating all the
CPU if the browser crashes before opening the pipe.

Also add an initial polling-connection phase to ClutterMozEmbed that
uses idle time to poll for the pipe-file existence. This finds the file
noticeably quicker than if we just rely on GFileMonitor, so spend an
initial 1 second polling in idle time (and after that, just rely on
GFileMonitor, and eventually the above connection timeout).