Fix extra motion event between acknowledgements
When we had a buffered motion event, we'd send it as soon as we got an
acknowledgement, but we forgot to then wait for an acknowledgement of
that event. That meant it was possible to send two motion events between
acknowledgements and end up starving drawing while scrolling. Now fixed.