[embed] [headless] Fix input handling
The condition parameter in GIOFunc is a bit-field, and can have multiple
values set. This fixes the handling when more than one bit is set.

Also, fix the situation where allocate was called with a negative value,
causing massive memory allocations (which worked most of the time
because they failed, but could cause problems).