[downloads] Reference the owning MozHeadless
While a download is active, it holds a reference on the MozHeadless that
the download came from. This stops the process from quitting if all
views disconnect before a download has completed.

Also, disable debugging spew as its frequency would block anything
running in the same terminal (for me at least). We know it works now
anyway.