Set the CLUTTER_TEXTURE_RGB_FLAG_BGR and _PREMULT flags for texture data
The texture data is coming from Cairo so it is actually
pre-multiplied. Without these flags Cogl will try to convert it using
a slow software path. We also need to specify these flags when
creating the empty data in the allocate function otherwise Cogl will
create the texture as RGB and try to convert the BGR data every time a
sub-region is updated.