[headless] Implement nsICertOverrideService to simplify bad cert handling
Every time the bad certificate handler in Mozilla is called, it checks
this override service and we always answer that we have an override so
the page will load. But we remember that we had a bad certificate on this
hostname/port, remember it, and add a flag to the security info whenever
the we are at this location. The UI can then inform the user in some way,
even while allowing them to see the web content.