Move responsibility for creating the plugin viewport window out to moz-headless
Because clutter-mozembed is in a seperate process to moz-headless the plugin
viewport window belonged to a different X client which caused problems when
the mozembed process was terminating because it would destroy the viewport
window from underneath all the plugins.

This simplifies a number of things in clutter-mozembed including no longer
needing to queue urls to open until we have setup the plugin viewport.