Replace ASCII pipes protocol with binary
Instead of communicating in human-readable ASCII, communicate in binary
instead. This was always intended, the ASCII was only there for
debugging. This results in a pretty noticeable speed-up (assuming it
isn't placebo effect), as we are no longer needlessly allocating and
freeing loads of memory and converting numbers to and from ASCII.

This also results in more shared code between embed/headless, but there
could still be more shared.