[downloads] Allow disabling the download file picker with a pref
If the boolean pref clutter_mozembed.disable_download_file_picker is
true then it will not prompt the user for a filename and instead just
download it directly.  The clutter_mozembed.download_directory string
pref controls where to put the file. If it is empty, the XDG download
dir is used, otherwise the value should be the name of a directory. If
the file already exists it will try to make a new filename by adding
-2 or -3 and so on until the file is available.