Use ACTOR_IS_MAPPED to determine when the plugin_viewport should be mapped
We used to use CLUTTER_ACTOR_IS_VISIBLE before actors supported a MAPPED
status but that's no longer the best way for us to know when mozembed actors
are mapped and visible on the stage (i.e. the MAPPED flag implies that all
ancestors are also MAPPED.)

This fixes a regression where plugin input was broken when switching between
multiple tabs.