[headless] Pass create_view copies of input/output
Although the current code won't crash, when the last view disconnects,
the input and output file properties become invalid. The last view
disconnecting doesn't necessarily mean the backend quits (in the case of
downloads, for example), and if we added code later that read either of
these properties at this time, we'd get an invalid pointer access.