[downloads] Throttle the progress reports
The OnProgressChange method gets called very frequently so if we try
to report everyone of them then it can starve the UI of the containing
browser. This patch prevents from sending an update if the last update
was less than 160ms ago. This is based on what happens in
nsBrowserStatusFilter.cpp in the Mozilla code.