[comms] Handle NULL pointers for strings
We'd crash on strlen (NULL) if we encountered a NULL string before,
which can sometimes happen for page titles (though I've never seen it
myself). Change the code to handle sending and receiving NULL for
strings (sends size=0, receiving end interprets that as a NULL string).

(Probably) Fixes Moblin bug #4951.