[embed] Catch when mapped memory has changed size for read-only case
Only the primary view ever _causes_ a resize, but the read-only previews
are still affected by the shared memory getting resized. If the resized
region was bigger, we probably just had a visual glitch w/ striping. But
when the new region had shrunk, we could get crashes. This was occurring
frequently as scrollbars appeared as the page reflowed, shrinking the web
region.

Fixes Moblin BZ 5130, 3818, and probably 5012