[embed] Trap X errors when syncing plugin viewport
When synchronising the plugin viewport position, we need to trap X
errors. The common crash we can get is if the backend disappears, the
window ID is no longer valid - on the one hand, it's still a bug that we
don't invalidate the handle when we detect the backend going down, but
X is asynchronous, so we need to trap this error too.