[prefs] Allow getting the default value for localized string prefs
Localized string prefs are retrieved using GetComplexValue with the
type set to nsIPrefLocalizedString. If there is a user setting it is
stored directly as a string otherwise the default value is a URL
pointing to a .properties file containing the value. The properties
file is usually a localized chrome URL so the value can change
depending on the locale.

This patch copies the GetDefaultFromPropertiesFile method from
nsPrefBranch which is used to fetch the default value for these
properties. The method is slightly modified to get the default value
using MHS calls rather than the private internal PREF_* API.

The localized string prefs are at least needed to get the
intl.accept_languages pref to work. That pref is used to set the HTTP
AcceptLanguages header depending on the user's locale.

Fixes Moblin bug #5812