[cookies] Allow newline characters in the cookie header
It seems that web servers use newlines instead of commas to separate
multiple cookies in a single header despite the spec saying that they
should be separated by commas. nsCookieService.cpp has this comment:

 ** Differences from RFC2109/2616 and explanations:

 ...

    2. We use CR | LF as cookie separators, not ',' per spec, since ',' is in
       common use inside values.

So it seems there is an alternative unofficial cookie standard. The
newline character is not accepted by g_ascii_isprint so any site that
was setting multiple cookies was failing. To fix this it now allows
whitespace in the header.

Fixes Moblin bug #5772