Enable Mozilla's private browsing mode when in mozheadless' private mode
This makes an implementation of nsIPrivateBrowsingService which will
report that browsing mode is enabled when the private property is set
on the ClutterMozEmbed.  This largely won't make much difference
because none of the services that are pertinent to private mode are
started anyway but it provides a convenient way for apps to detect
private mode.