[headless] Fix mistakenly showing broken security
If we had a secure site with insecure elements (for example, loading a
mail in Google Mail that accesses external images), we would show broken
security. Fixed to look at the URI on security-change and check if the
change applies to the page or an element on the page. Requires latest
mozilla-headless.

Fixes Moblin bug #6300.

Also some small changes with resizing that may help with bug #7436.